move tests to a test folder
