+++ effect-system/tests/effects.test.ts
@@ -3,103 +3,25 @@ import { assertEquals } from "@std/assert/equals";
 import { assertGreaterOrEqual } from "@std/assert/greater-or-equal";
 import { assertLessOrEqual } from "@std/assert/less-or-equal";
 import { afterEach, beforeEach, describe, test } from "@std/testing/bdd";
-import { createEffect, handlerFor } from "./effects.ts";
-import { addHandlers, Handler, handlerScopes } from "./handlers.ts";
-import { HandlerScope, id } from "./mod.ts";
-
-/**
- * Console
- */
-
-interface ConsoleOps {
-  log: (message: string) => void;
-}
-
-const Console = { log: createEffect<ConsoleOps["log"]>("console/log") };
-
-const logs: string[] = [];
-
-const handleConsole = handlerFor(Console.log, (message: string) => {
-  logs.push(message);
-});
-
-/**
- * State
- */
-
-interface StateOps<S> {
-  get: () => S;
-  set: (state: S) => void;
-  update: (updater: (old: S) => S) => void;
-}
-
-const createState = <S>(initialState: S) => {
-  const get = createEffect<StateOps<S>["get"]>("state/get");
-  const set = createEffect<StateOps<S>["set"]>("state/set");
-  const update = createEffect<StateOps<S>["update"]>("state/update");
-
-  let state = initialState;
-
-  return {
-    get,
-    set,
-    update,
-    handlers: [
-      handlerFor(get, () => state),
-      handlerFor(set, (newState) => {
-        state = newState;
-      }),
-      handlerFor(update, (updater) => {
-        state = updater(state);
-      }),
-    ],
-  };
-};
-
-const state = createState(0);
-
-/**
- * Random
- */
-
-interface RandomOps {
-  random: () => number;
-}
-
-const random = createEffect<RandomOps["random"]>("random");
-
-const handleRandom = handlerFor(random, () => Math.random());
-
-/**
- * IO
- */
-
-interface IO {
-  readFile: (path: string) => string;
-  writeFile: (path: string, data: string) => void;
-  transformFile: (
-    options: { path: string; data: string },
-  ) => { path: string; data: string };
-}
-
-const io = {
-  readFile: createEffect<IO["readFile"]>("io/read"),
-  writeFile: createEffect<IO["writeFile"]>("io/write"),
-  transformFile: createEffect<
-    (options: { path: string; data: string }) => { path: string; data: string }
-  >("io/transform"),
-};
-
-const handleIoReadTXT = handlerFor(io.readFile, (path: string) => {
-  if (path.endsWith(".txt")) {
-    return "txt content";
-  }
-  return Handler.continue(path);
-});
-
-const handleIOReadBase = handlerFor(io.readFile, () => {
-  return "file content";
-});
+import { handlerFor } from "../effects.ts";
+import { addHandlers, Handler, handlerScopes } from "../handlers.ts";
+import { HandlerScope, id } from "../mod.ts";
+import {
+  Console,
+  createState,
+  handleConsole,
+  handleIOReadBase,
+  handleIoReadTXT,
+  handleRandom,
+  io,
+  logs,
+  random,
+} from "./setup.ts";
+import {
+  MissingHandlerScopeError,
+  MissingTerminalHandlerError,
+  UnhandledEffectError,
+} from "../errors.Ts";
 
 /**
  * Tests
@@ -122,8 +44,7 @@ describe("effect system", () => {
       await random();
       unreachable();
     } catch (error) {
-      assertInstanceOf(error, Error);
-      assertEquals(error.message, 'Unhandled effect "random"');
+      assertInstanceOf(error, UnhandledEffectError);
     }
   });
 
@@ -154,11 +75,7 @@ describe("effect system", () => {
       await io.readFile("style.css");
       unreachable();
     } catch (error) {
-      assertInstanceOf(error, Error);
-      assertEquals(
-        error.message,
-        'Handling effect "io/read" returned `Continue`. Make sure the handlers sequence contains a terminal handler',
-      );
+      assertInstanceOf(error, MissingTerminalHandlerError);
     }
   });
 
@@ -186,11 +103,7 @@ describe("effect system", () => {
       addHandlers([handleIoReadTXT]);
       unreachable();
     } catch (error) {
-      assertInstanceOf(error, Error);
-      assertEquals(
-        error.message,
-        '"addHandlers" called outside of an effect scope',
-      );
+      assertInstanceOf(error, MissingHandlerScopeError);
     }
   });
 
@@ -206,11 +119,7 @@ describe("effect system", () => {
       await io.readFile("note.txt");
       unreachable();
     } catch (error) {
-      assertInstanceOf(error, Error);
-      assertEquals(
-        error.message,
-        'Effect "io/read" running outside of a HandlerScope',
-      );
+      assertInstanceOf(error, MissingHandlerScopeError);
     }
   });
 
@@ -267,6 +176,8 @@ describe("effect system", () => {
   });
 
   test("observation", async () => {
+    const state = createState(0);
+
     // the countWrite handler only observe the io/write effect and does something orthogonal
     const countWrites = handlerFor(io.writeFile, async (path, data) => {
       await state.update((old) => old + 1);
